email_template_management.php
This is the email template management code. which creates email templates in backend we can create different email templates and call them in the function where we want to send an email.

We can call the content if the user uploads any in the content area. Then we can send the template by checking if there is any content in the fields.otherwise we can send default email.